Split manifest test code into helper file
